requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
httpx==0.28.1
python-dotenv==1.1.1
redis==6.4.0
psycopg[binary]==3.2.9
transformers==4.55.4
torch==2.8.0
sentence-transformers==5.1.0
faiss-cpu==1.12.0
pypdf==6.10.2
